Changed how database values are casted in.

+++ cicm_web/Controllers/CompanyController.cs
@@ -68,7 +68,7 @@ namespace cicm_web.Controllers
             return View(company);
         }
 
-        public IActionResult ByCountry(int id)
+        public IActionResult ByCountry(ushort id)
         {
             Iso3166 iso3166 = Program.Database.Operations.GetIso3166(id);
